refactor(object): split hasOwn helpers
