+++ crates/perry-runtime/src/object/mod.rs
@@ -34,6 +34,7 @@ mod delete_rest;
 mod field_get_set;
 mod field_set_by_name;
 mod global_this;
+mod has_own_helpers;
 mod instanceof;
 mod native_call_method;
 mod native_module;
